discord_member_roles_fetcher/requirements.txt
discord==2.3.2
pandas==2.2.2
google-cloud-bigquery==3.20.1
google-cloud-secret-manager==2.20.0
python-dotenv==1.0.0